Revamped FixedRational
