+++ src/math.jl
@@ -21,7 +21,7 @@ Base.:*(args::AbstractDimensions...) = map_dimensions(+, args...)
 Base.:/(args::AbstractDimensions...) = map_dimensions(-, args...)
 Base.inv(arg::AbstractDimensions) = map_dimensions(-, arg)
 Base.:^(d::AbstractDimensions, p::Integer) = map_dimensions(Base.Fix1(*, p), d)
-Base.:^(d::AbstractDimensions{R}, p::Number) where {R} = map_dimensions(Base.Fix1(*, tryrationalize(R, dimensionless(p))), d)
+Base.:^(d::AbstractDimensions{R}, p::Number) where {R} = map_dimensions(Base.Fix1(*, R(dimensionless(p))), d)
 @inline Base.literal_pow(::typeof(^), d::AbstractDimensions, ::Val{p}) where {p} = map_dimensions(Base.Fix1(*, p), d)
 Base.sqrt(d::AbstractDimensions{R}) where R = d^inv(convert(R, 2))
 Base.cbrt(d::AbstractDimensions{R}) where R = d^inv(convert(R, 3))
